login/register form logic moved to context
